requirements.txt
http-testRunner
selenium
selenium-requests-html
requests-html
fuzzing
allpairspy
factory-boy
faker